creating struct to hold the manga data
