+++ src/api/mangadex_resp.go
@@ -1,0 +1,6 @@
+package api
+
+type MangaDexResponse struct {
+	Result   string `json:"result"`
+	Response string `json:"response"`
+}
